services/log-aggregator/requirements.txt
confluent-kafka==2.4.0
clickhouse-driver==0.2.7
pika==1.3.2
python-dotenv==1.0.1
